initial timechart dataprocessor setup
