+++ pkg/segment/query/processor/queryprocessor.go
@@ -76,7 +76,7 @@ func NewQueryProcessor(firstAgg *structs.QueryAggregators, queryInfo *query.Quer
 
 	dataProcessors := make([]*DataProcessor, 0)
 	for curAgg := firstProcessorAgg; curAgg != nil; curAgg = curAgg.Next {
-		dataProcessor := asDataProcessor(curAgg)
+		dataProcessor := asDataProcessor(curAgg, queryInfo)
 		if dataProcessor == nil {
 			break
 		}
@@ -128,7 +128,7 @@ func newQueryProcessorHelper(queryType structs.QueryType, input streamer,
 	}, nil
 }
 
-func asDataProcessor(queryAgg *structs.QueryAggregators) *DataProcessor {
+func asDataProcessor(queryAgg *structs.QueryAggregators, queryInfo *query.QueryInformation) *DataProcessor {
 	if queryAgg == nil {
 		return nil
 	}
@@ -159,6 +159,8 @@ func asDataProcessor(queryAgg *structs.QueryAggregators) *DataProcessor {
 		return NewRexDP(queryAgg.RexExpr)
 	} else if queryAgg.SortExpr != nil {
 		return NewSortDP(queryAgg.SortExpr)
+	} else if queryAgg.TimeHistogram != nil && queryAgg.TimeHistogram.Timechart != nil {
+		return NewTimechartDP(queryAgg, queryInfo.GetQueryRange(), queryInfo.GetQid())
 	} else if queryAgg.GroupByRequest != nil {
 		queryAgg.StatsExpr = &structs.StatsExpr{GroupByRequest: queryAgg.GroupByRequest}
 		return NewStatsDP(queryAgg.StatsExpr)
@@ -171,8 +173,6 @@ func asDataProcessor(queryAgg *structs.QueryAggregators) *DataProcessor {
 		return NewStreamstatsDP(queryAgg.StreamstatsExpr)
 	} else if queryAgg.TailExpr != nil {
 		return NewTailDP(queryAgg.TailExpr)
-	} else if queryAgg.TimechartExpr != nil {
-		return NewTimechartDP(queryAgg.TimechartExpr)
 	} else if queryAgg.TopExpr != nil {
 		return NewTopDP(queryAgg.TopExpr)
 	} else if queryAgg.TransactionExpr != nil {
