+++ pkg/segment/query/processor/timechartcommand.go
@@ -18,12 +18,54 @@
 package processor
 
 import (
+	"github.com/siglens/siglens/pkg/common/dtypeutils"
+	"github.com/siglens/siglens/pkg/segment/aggregations"
 	"github.com/siglens/siglens/pkg/segment/query/iqr"
 	"github.com/siglens/siglens/pkg/segment/structs"
+	"github.com/siglens/siglens/pkg/segment/utils"
 )
 
+type timechartOptions struct {
+	timeBucket     *structs.TimeBucket
+	timeChartExpr  *structs.TimechartExpr
+	groupByRequest *structs.GroupByRequest
+}
+
 type timechartProcessor struct {
-	options *structs.TimechartExpr
+	options   *timechartOptions
+	spanError error
+}
+
+func NewTimechartProcessor(aggs *structs.QueryAggregators, timeRange *dtypeutils.TimeRange, qid uint64) *timechartProcessor {
+	if aggs.TimeHistogram == nil || aggs.TimeHistogram.Timechart == nil {
+		return &timechartProcessor{options: nil}
+	}
+
+	if aggs.GroupByRequest != nil {
+		aggs.GroupByRequest.BucketCount = int(utils.QUERY_MAX_BUCKETS)
+	}
+
+	processor := &timechartProcessor{}
+
+	if aggs.TimeHistogram.Timechart.BinOptions != nil &&
+		aggs.TimeHistogram.Timechart.BinOptions.SpanOptions != nil &&
+		aggs.TimeHistogram.Timechart.BinOptions.SpanOptions.DefaultSettings {
+		spanOptions, err := structs.GetDefaultTimechartSpanOptions(timeRange.StartEpochMs, timeRange.EndEpochMs, qid)
+		if err != nil {
+			processor.spanError = err
+			return processor
+		}
+		aggs.TimeHistogram.Timechart.BinOptions.SpanOptions = spanOptions
+		aggs.TimeHistogram.IntervalMillis = aggregations.GetIntervalInMillis(spanOptions.SpanLength.Num, spanOptions.SpanLength.TimeScalr)
+	}
+
+	processor.options = &timechartOptions{
+		timeBucket:     aggs.TimeHistogram,
+		timeChartExpr:  aggs.TimeHistogram.Timechart,
+		groupByRequest: aggs.GroupByRequest,
+	}
+
+	return processor
 }
 
 func (p *timechartProcessor) Process(iqr *iqr.IQR) (*iqr.IQR, error) {
